pyproject.toml
[tool.poetry]
name = "django-rest-framework-async"
version = "0.1.0"
description = ""
authors = ["frownyface"]
readme = "README.md"
packages = [{include = "drfa"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.1"
pytest-asyncio = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
